fix(rest-service): make shutdown sender available
We might need to send the shutdown signal from outside the service, thus
it's a good idea to make it available. The first place where we see the
need is in the tests as we need to gracefully shutdown the server after
every test to avoid trying to bind on the same port over and over.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>