fix(task-processor): change event on simulation error
When a simulation error happens in the execute_task flow, we might have
already uploaded the payload to solana and thus incurred costs. As costs
can only be passed in the MESSAGE_EXECUTED event, we need to use it to
report back to Axelar. This is a temporary fix until the Amplifier API
is updated to accept all txs in the flow, when this happens, we'll also
need to keep track of all the transactions that went beyond the
simulation step.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>