feat: fee estimation before execution
This patch adds a gas estimation module which is supposed to work with a
local node with a fork of the actual Solana state. The reason this is
necessary is because due to transaction size limitations on Solana and
Axelar minimum requirements, we need to perform Axelar tasks through
multiple Solana transactions. Before performing the task though, we need
to know if the available gas is enough to cover the costs of all
transactions required for that: initializing the message payload,
uploading the payload on-chain, committing the payload, invoking the
destination program with the payload account, and then closing the
payload account.

Normally the estimation is done through the `simulate_transaction` call
on the Solana RpcClient, but since multiple transactions are required
for an Axelar task, the transactions that need accounts created by
earlier transactions would fail, since the transactions that create
them are not really broadcast on the real chain and these accounts
are thus not created on-chain.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>